Update to accept title option and Twig templates
The title is one of the most important values of a template. The user
need to be able to set it. The command prompts them for it, just in case
he or she forgets to specify it on the command line.

When Twig specifies a value for output, the default output is just a
name between two curly braces. It's a better experience value a template
name doesn't look like a variable.